Shrink varTableStamp to a byte
This stamp is used to guard atomic access to the varTable field
when it needs to be grown to accommodate more values. The logic
expects that anyone performing such an update will increment the
stamp while updating the the table reference, backing off if the
stamp has changed before finally assigning the new table (allowing
concurrent writers to all attempt growth without a hard lock, with
the first-past-the-post thread winning).

We probably don't need 32 bit for this, since that would
accommodate an absurd 2-4 billion concurrent mutators. Byte range
may be too small, but still handles at least 64 concurrent mutators
with non successfully growing the table large enough for any of the
others to proceed.

This relates to object header size reductions detailed in recent
issues like jruby/jruby#9090.

+++ core/src/main/java/org/jruby/runtime/ivars/AtomicVariableTable.java
@@ -23,7 +23,7 @@ public class AtomicVariableTable {
      */
     public static void setVariableAtomic(RubyBasicObject self, RubyClass realClass, boolean fullFence, int index, Object value) {
         while (true) {
-            int currentStamp = self.varTableStamp;
+            byte currentStamp = self.varTableStamp;
             // spin-wait if odd
             if((currentStamp & 0x01) != 0)
                 continue;
@@ -51,7 +51,7 @@ public class AtomicVariableTable {
      * @param value the variable's value
      * @return whether the update was successful, for CAS retrying
      */
-    private static boolean createTableAtomic(RubyBasicObject self, int currentStamp, RubyClass realClass, Object[] currentTable, int index, Object value) {
+    private static boolean createTableAtomic(RubyBasicObject self, byte currentStamp, RubyClass realClass, Object[] currentTable, int index, Object value) {
         // try to acquire exclusive access to the varTable field
         if (!VariableAccessor.STAMP_HANDLE.compareAndSet(self, currentStamp, ++currentStamp)) {
             return false;
@@ -73,7 +73,7 @@ public class AtomicVariableTable {
         VariableAccessor.VAR_TABLE_HANDLE.setRelease(self, newTable);
 
         // release exclusive access
-        self.varTableStamp = currentStamp + 1;
+        self.varTableStamp = (byte) (currentStamp + 1);
 
         return true;
     }
@@ -88,7 +88,7 @@ public class AtomicVariableTable {
      * @param value the variable's value
      * @return whether the update was successful, for CAS retrying
      */
-    private static boolean updateTableAtomic(RubyBasicObject self, int currentStamp, Object[] currentTable, int index, Object value) {
+    private static boolean updateTableAtomic(RubyBasicObject self, byte currentStamp, Object[] currentTable, int index, Object value) {
         // shared access to varTable field.
         currentTable[index] = value;
         VarHandle.fullFence();
